+++ Recepies/Services/Network/DTO/CategoryDTO/CategoryDTO.swift
@@ -1,0 +1,12 @@
+// CategoryDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+// MARK: - CategoryDTO
+
+/// DTO for search by category
+struct CategoryDTO: Codable {
+    /// Array of recipesDTO containers
+    let hits: [HitDTO]
+}
