+++ Recepies/Services/Network/DTO/CategoryDTO/HitDTO.swift
@@ -1,0 +1,10 @@
+// HitDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+// MARK: - Hit
+
+/// RecipeDTO container
+struct HitDTO: Codable {
+    /// Recipe
+    let recipe: RecipeDTO
+}
