+++ Recepies/Services/Network/DTO/DetailDTO/InfoDTO.swift
@@ -1,0 +1,8 @@
+// InfoDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+/// Structure o Quantity of nutritional substance
+struct InfoDTO: Codable {
+    /// Amount of PFC components in gramms
+    let quantity: Double
+}
