+++ Recepies/Services/Network/DTO/DetailDTO/RecipeDTOContainer.swift
@@ -1,0 +1,8 @@
+// RecipeDTOContainer.swift
+// Copyright © RoadMap. All rights reserved.
+
+/// Structure for decoding responses for the server
+struct RecipeDTOContainer: Codable {
+    /// Detail recipe
+    let recipe: RecipeDetailDTO
+}
