+++ Recepies/Services/Network/DTO/DetailDTO/RecipeDetailResponseDTO.swift
@@ -1,0 +1,9 @@
+// RecipeDetailResponseDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// General data container for Detailed recipe fromserver
+struct RecipeDetailResponseDTO: Codable {
+    let hits: [RecipeDTOContainer]
+}
