+++ Recepies/Services/Network/DTO/DetailDTO/TotalDTO.swift
@@ -1,0 +1,8 @@
+// TotalDTO.swift
+// Copyright © RoadMap. All rights reserved.
+
+/// TotalDTO
+struct TotalDTO: Codable {
+    /// Amount of PFC components in gramms
+    let quantity: Double
+}
